pyproject.toml
[project]
name = "uppetit-promo-bot"
version = "1.0.0"
description = "Telegram bot for UPPETIT promo QR codes distribution"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.13.1",
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "asyncpg>=0.30.0",
    "aiosqlite>=0.20.0",
    "qrcode[pil]>=8.0",
    "python-dotenv>=1.0.1",
    "structlog>=24.4.0",
    "pytz>=2024.2",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_backend"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*", "tools*"]
